requirements.txt
asgiref==3.8.1
djangorestframework==3.15.2
Django==5.1.3
psycopg[binary]==3.2.3

# Asynchronous Support & Async Django Rest Framework
uvicorn[standard]==0.32.0

# pytest in requirements, not dev, because it is django-test project
pytest-django==4.9.0
